Update and rename SendSmsController to app/controller/SendSmsController
In this Laravel code, the Http facade is used to send an HTTP POST request to the specified URL ($server). The request body is serialized as JSON using the json_encode function. The withHeaders method is used to set the Content-Type header to 'application/json'. The response body is obtained using the body method, and the HTTP status code is obtained using the status method. Finally, the response is printed using print_r